Remove potential deadlock from synchronizer.
Previously we did not hold the condition variable lock before notifying
from either the worker or the manager thread. Although the condition
itself is an atomic variable, this could still result in a missed
notification. Consider the following scenario:

        Worker:                     Manager:

        acquires lock
        checks condition
        (descheduled here)
                                    changes condition
                                    notifies workers
        releases lock               (descheduled here)
        sleeps

The worker would miss the notification here. At this point, we probably
don't need atomic conditions, and everything would be simpler if we used
boost::shared_mutex.