Allow use of any container for BitVector.
Previously, the template specialization of the heuristic feasibility
operator for a dynamic_bitset prevented code from compiling when other
containers were used as BitVector. Now we use enable_if and some
template magic to only compile the specialization if it is necessary.
Tag dispatching may be better suited for this.